+++ src/main/kotlin/com/books/backend/models/entity/Book.kt
@@ -23,8 +23,8 @@ data class Book(
     val title: String? = null,
     val isbn: String? = null,
     val publicDate: Timestamp? = null,
-    val publisherId: Int? = null,
-    val genreId: Int? = null,
+    val publisherId: Long? = null,
+    val genreId: Long? = null,
     val summary: String? = null,
-    val pageCount: Int? = null,
+    val pageCount: Long? = null,
 ): Serializable
